Move route blinding construction to router
